OpenXR - Remove GLES3 dependency, start migrating to C
