+++ app/src/main/cpp/lorie/xr/framebuffer.h
@@ -1,44 +1,42 @@
 #pragma once
 
-#ifdef ANDROID
-#include <jni.h>
+//TODO:remove once base.h is in C
 #define XR_USE_PLATFORM_ANDROID 1
 #define XR_USE_GRAPHICS_API_OPENGL_ES 1
 #include <EGL/egl.h>
 #include <EGL/eglext.h>
-#endif
+#include <android/log.h>
+#include <jni.h>
+#define ALOGE(...) __android_log_print(ANDROID_LOG_ERROR, "OpenXR", __VA_ARGS__);
+#define ALOGV(...) __android_log_print(ANDROID_LOG_VERBOSE, "OpenXR", __VA_ARGS__);
+#define GL(func) func;
+#define OXR(func) func;
 
 #include <openxr/openxr.h>
 #include <openxr/openxr_platform.h>
+#include <stdbool.h>
 
-class Framebuffer
-{
-public:
-  bool Create(XrSession session, int width, int height);
-  void Destroy();
+struct XrFramebuffer {
+    int Width;
+    int Height;
+    bool Acquired;
+    XrSwapchain Handle;
 
-  int GetWidth() { return m_width; }
-  int GetHeight() { return m_height; }
-  XrSwapchain GetHandle() { return m_handle; }
+    uint32_t SwapchainIndex;
+    uint32_t SwapchainLength;
+    void* SwapchainImage;
 
-  void Acquire();
-  void Release();
-  void SetCurrent();
-
-private:
-#if XR_USE_GRAPHICS_API_OPENGL_ES
-  bool CreateGL(XrSession session, int width, int height);
-#endif
-
-  int m_width;
-  int m_height;
-  bool m_acquired;
-  XrSwapchain m_handle;
-
-  uint32_t m_swapchain_index;
-  uint32_t m_swapchain_length;
-  void* m_swapchain_image;
-
-  unsigned int* m_gl_depth_buffers;
-  unsigned int* m_gl_frame_buffers;
+    unsigned int* GLDepthBuffers;
+    unsigned int* GLFrameBuffers;
 };
+
+bool XrFramebufferCreate(struct XrFramebuffer *framebuffer, XrSession session, int width, int height);
+void XrFramebufferDestroy(struct XrFramebuffer *framebuffer);
+
+void XrFramebufferAcquire(struct XrFramebuffer *framebuffer);
+void XrFramebufferRelease(struct XrFramebuffer *framebuffer);
+void XrFramebufferSetCurrent(struct XrFramebuffer *framebuffer);
+
+#if XR_USE_GRAPHICS_API_OPENGL_ES
+bool XrFramebufferCreateGL(struct XrFramebuffer *framebuffer, XrSession session, int width, int height);
+#endif
