+++ app/src/main/cpp/lorie/xr/input.cpp
@@ -1,5 +1,8 @@
 #include "input.h"
-#include "math.h"
+
+extern "C" {
+    #include "math.h"
+}
 
 #include <cstring>
