+++ app/src/main/cpp/recipes/xserver.cmake
@@ -268,10 +268,10 @@ add_library(Xlorie SHARED
         "libxcvt/lib/libxcvt.c"
         "lorie/shm/shmem.c"
         "lorie/xr/base.cpp"
-        "lorie/xr/framebuffer.cpp"
+        "lorie/xr/framebuffer.c"
         "lorie/xr/input.cpp"
         "lorie/xr/main.cpp"
-        "lorie/xr/math.cpp"
+        "lorie/xr/math.c"
         "lorie/xr/renderer.cpp"
         "lorie/android.c"
         "lorie/InitOutput.c"
